Add security checks when using Nginx upload-module
You must now specify in the configuration that you are using the Nginx
upload-module (it is no more automatically detected).
The store of the upload-module must also be specified to deny moving
files outside of the store (for example, if a malicious user forge a
POST request with `myfile.path=/etc/password`).