Add support for a basic password protection for uploads (#2)
If FLASKUP_UPLOAD_PASSWORDS is defined (as a list of valid passwords),
Flaskup will ask users to provide a password from this list when he
upload a file. If the password is not known the upload is denied.
If FLASKUP_UPLOAD_PASSWORDS is not defined (or the list is empty),
Flaskup acts as normal: no passwords needed.